refactor: Modularize dayMap code
